requirements.txt
click>=7.0
pyserial
pymodbus
